backend/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyotp==2.9.0
qrcode==7.4.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Background Tasks
celery==5.3.6
redis==5.0.1

# File Processing
python-magic==0.4.27
Pillow==10.2.0

# OCR
pytesseract==0.3.10
pdf2image==1.17.0
PyPDF2==3.0.1

# Encryption
cryptography==42.0.2

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
uuid6==2024.1.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4

# Development
black==24.1.1
isort==5.13.2
